+++ R/pacmap.R
@@ -77,6 +77,8 @@
 #' @param pacmap_params Named list. PaCMAP algorithm parameters, see
 #' [params_pacmap()].
 #' @param seed Integer. Random seed for reproducibility. Defaults to `42L`.
+#' @param use_high_precision Optional boolean. Gives fine-grained control over
+#' `fp32` vs `fp64` usage.
 #' @param .verbose Logical. Controls verbosity. Defaults to `TRUE`.
 #'
 #' @return A numerical matrix with dimensions samples x n_dim containing
@@ -99,6 +101,7 @@ pacmap <- function(
   nn_params = params_nn(),
   pacmap_params = params_pacmap(),
   seed = 42L,
+  use_high_precision = NULL,
   .verbose = TRUE
 ) {
   if (is.data.frame(data)) {
@@ -120,6 +123,7 @@ pacmap <- function(
   checkmate::assert_int(n_dim, lower = 1, upper = ncol(data))
   checkmate::qassert(k, "I1[2,)")
   checkmate::qassert(seed, "I1")
+  checkmate::qassert(use_high_precision, c("0", "B1"))
   checkmate::qassert(.verbose, c("B1", "I1[0, 2]"))
 
   final_pacmap_params <- .prepare_pacmap_params(
@@ -142,6 +146,7 @@ pacmap <- function(
           k = k,
           pacmap_params = final_pacmap_params,
           seed = seed,
+          use_high_precision = use_high_precision,
           verbose = parse_verbosity(.verbose)
         )
       },
@@ -158,6 +163,7 @@ pacmap <- function(
           k = k,
           pacmap_params = final_pacmap_params,
           seed = seed,
+          use_high_precision = use_high_precision,
           verbose = parse_verbosity(.verbose)
         )
       },