+++ src/rust/src/umap.rs
@@ -2,11 +2,14 @@
 
 #![warn(missing_docs)]
 
+use ann_search_rs::cpu::hnsw::{HnswIndex, HnswState};
+use ann_search_rs::cpu::nndescent::{ApplySortedUpdates, NNDescent, NNDescentQuery};
 use bixverse_rs::prelude::IntoExtendrErr;
 use extendr_api::{List, Robj};
 use faer::{Mat, MatRef};
 use manifolds_rs::prelude::*;
 use manifolds_rs::*;
+use rand_distr::{Distribution, StandardNormal};
 use std::collections::HashMap;
 
 use crate::utils::get_params_nn_manifolds;
@@ -20,14 +23,14 @@ use crate::utils::get_params_nn_manifolds;
 /// Internal representation of various parameters needed for the UMAP
 /// implementation in `manifolds-rs`.
 #[derive(Debug)]
-pub struct InternalUmapParams {
+pub struct InternalUmapParams<T> {
     /// Which of the approximate nearest neighbour searches to use.
     pub knn_method: String,
     /// The nearest neighbour parameters that are forwarded to the approximate
     /// nearest neighbour methods.
-    pub param_knn: NearestNeighbourParams<f32>,
+    pub param_knn: NearestNeighbourParams<T>,
     /// The UMAP graph generation parameters.
-    pub umap_graph: UmapGraphParams<f32>,
+    pub umap_graph: UmapGraphParams<T>,
     /// Which initialisation to use. One of `"spectral"`, `"pca"`, or
     /// `"random"`.
     pub init: String,
@@ -38,10 +41,13 @@ pub struct InternalUmapParams {
     /// `"adam_parallel"`.
     pub optimiser: String,
     /// The UMAP optimisation parameters.
-    pub param_optimiser: UmapOptimParams<f32>,
+    pub param_optimiser: UmapOptimParams<T>,
 }
 
-impl InternalUmapParams {
+impl<T> InternalUmapParams<T>
+where
+    T: ManifoldsFloat,
+{
     /// Generate the UMAP parameters from an R list
     ///
     /// ### Params
@@ -55,8 +61,8 @@ impl InternalUmapParams {
     /// The `BixverseUmapParams`.
     pub fn from_r_list(
         r_list: List,
-        min_dist: f32,
-        spread: f32,
+        min_dist: f64,
+        spread: f64,
     ) -> Result<Self, extendr_api::Error> {
         let nn_params = get_params_nn_manifolds(r_list.clone())?;
         let umap_graph_params = get_params_umap_graph(r_list.clone())?;
@@ -111,23 +117,29 @@ impl InternalUmapParams {
 /// ### Returns
 ///
 /// The `UmapGraphParams` with sensible defaults if not found in the list.
-pub fn get_params_umap_graph(r_list: List) -> Result<UmapGraphParams<f32>, extendr_api::Error> {
+pub fn get_params_umap_graph<T>(r_list: List) -> Result<UmapGraphParams<T>, extendr_api::Error>
+where
+    T: ManifoldsFloat,
+{
     let graph_params: HashMap<&str, Robj> = r_list.try_into()?;
 
-    let mix_weight = graph_params
+    let mix_weight: T = graph_params
         .get("mix_weight")
         .and_then(|v| v.as_real())
-        .unwrap_or(1.0) as f32;
+        .map(|v| T::from_f64(v).unwrap())
+        .unwrap_or(T::from_f64(1.0).unwrap());
 
-    let local_connectivity = graph_params
+    let local_connectivity: T = graph_params
         .get("local_connectivity")
         .and_then(|v| v.as_real())
-        .unwrap_or(1.0) as f32;
+        .map(|v| T::from_f64(v).unwrap())
+        .unwrap_or(T::from_f64(1.0).unwrap());
 
     let bandwidth = graph_params
         .get("bandwidth")
         .and_then(|v| v.as_real())
-        .unwrap_or(1e-5) as f32;
+        .map(|v| T::from_f64(v).unwrap())
+        .unwrap_or(T::from_f64(1e-5).unwrap());
 
     Ok(UmapGraphParams {
         bandwidth,
@@ -150,17 +162,23 @@ pub fn get_params_umap_graph(r_list: List) -> Result<UmapGraphParams<f32>, exten
 /// ### Returns
 ///
 /// The `UmapOptimParams` with sensible defaults if not found in the list.
-fn get_params_umap_optim(
+fn get_params_umap_optim<T>(
     r_list: List,
-    min_dist: f32,
-    spread: f32,
-) -> Result<UmapOptimParams<f32>, extendr_api::Error> {
+    min_dist: f64,
+    spread: f64,
+) -> Result<UmapOptimParams<T>, extendr_api::Error>
+where
+    T: ManifoldsFloat,
+{
     let optim_params: HashMap<&str, Robj> = r_list.try_into()?;
+
+    let min_dist = T::from_f64(min_dist).unwrap();
+    let spread = T::from_f64(spread).unwrap();
 
     let lr = optim_params
         .get("lr")
         .and_then(|v| v.as_real())
-        .map(|v| v as f32);
+        .map(|v| T::from_f64(v).unwrap());
 
     let n_epochs = optim_params
         .get("n_epochs")
@@ -175,7 +193,7 @@ fn get_params_umap_optim(
     let gamma = optim_params
         .get("gamma")
         .and_then(|v| v.as_real())
-        .map(|v| v as f32);
+        .map(|v| T::from(v).unwrap());
 
     Ok(UmapOptimParams::from_min_dist_spread(
         min_dist,
@@ -217,21 +235,27 @@ fn get_params_umap_optim(
 ///
 /// Returns the UMAP embeddings as matrix.
 #[allow(clippy::too_many_arguments)]
-pub fn umap_manifold(
-    data: MatRef<f32>,
-    pre_computed_knn: PreComputedKnn<f32>,
+pub fn umap_manifold<T>(
+    data: MatRef<T>,
+    pre_computed_knn: PreComputedKnn<T>,
     n_dim: usize,
     k: usize,
-    min_dist: f32,
-    spread: f32,
+    min_dist: f64,
+    spread: f64,
     umap_params: List,
     seed: usize,
     verbose: usize,
-) -> Result<Mat<f32>, extendr_api::Error> {
+) -> Result<Mat<T>, extendr_api::Error>
+where
+    T: ManifoldsFloat,
+    HnswIndex<T>: HnswState<T>,
+    NNDescent<T>: ApplySortedUpdates<T> + NNDescentQuery<T>,
+    StandardNormal: Distribution<T>,
+{
     let internal = InternalUmapParams::from_r_list(umap_params, min_dist, spread)?;
 
     let init_range = if internal.init == "pca" {
-        Some(1e-4)
+        Some(T::from_f64(1.0).unwrap())
     } else {
         None
     };